Add support for application-level connection closing.
Many sorts of network connections ought to be closed properly when they
are dropped. We now provide a way for pool holders to close all pooled
connections, for example on application exit. Ideally we'd be able to do
this implicitly, but we don't want to force all callers to hold the
connection pool in a context manager.